Restructure Python v3 (Part 5) - Rename files
