+++ deepgram/__init__.py
@@ -2,16 +2,16 @@
 __version__ = '0.0.0'
 
 # entry point for the deepgram python sdk
-from .deepgram_client import DeepgramClient
-from .deepgram_client_options import DeepgramClientOptions
+from .client import DeepgramClient
+from .options import DeepgramClientOptions
 from .errors import DeepgramError, DeepgramApiError, DeepgramUnknownApiError, DeepgramUnknownError
 
 # live
-from .clients.live.transcription_options import LiveOptions
+from .clients.live.options import LiveOptions
 from .clients.live.enums import LiveTranscriptionEvents
 
 # prerecorded
-from .clients.prerecorded.transcription_options import PrerecordedOptions
+from .clients.prerecorded.options import PrerecordedOptions
 
 # manage
-from .clients.manage.manage_client import ManageClient
+from .clients.manage.client import ManageClient
